# remove HEAD requests