closes #17 

cc @kylebarron 🙏