@vincentsarago 2025-02-27
```
# Before
{
  "HEAD": {
    "count": 1
  },
  "GET": {
    "count": 2,
    "bytes": 403064,
    "ranges": [
      "0-32768",
      "3622-253242",
      "460080-580753"
    ]
  },
  "Timing": 2.0073082447052,
  "cprofile": [
    "   ncalls  tottime  percall  cumtime  percall filename:lineno(function)",
    "        1    0.791    0.791    0.791    0.791 {get_ranges}",
    "        1    0.596    0.596    0.596    0.596 {head}",
    "        1    0.516    0.516    0.516    0.516 {get_range}",
    "        1    0.027    0.027    0.821    0.821 reader.py:86(read)",
    "        1    0.022    0.022    1.152    1.152 __init__.py:99(open)",
    "        1    0.006    0.006    0.006    0.006 {built-in method from_url}",
    "        1    0.005    0.005    0.006    0.006 warp.py:453(calculate_default_transform)",
    "        1    0.005    0.005    0.005    0.005 warp.py:113(transform_bounds)",
    "        8    0.004    0.000    0.004    0.000 {method 'execute' of 'sqlite3.Connection' objects}",
    "      156    0.002    0.000    0.004    0.000 __init__.py:298(__init__)",
    "        1    0.001    0.001    0.001    0.001 core.py:3521(__setmask__)",
    "        1    0.001    0.001    0.009    0.009 rasterio.py:94(__attrs_post_init__)",
    "        1    0.001    0.001    0.792    0.792 base.py:204(get_byte_ranges)",
    "      148    0.001    0.000    0.009    0.000 base.py:161(read)"
  ]
}

# This PR 
{
  "HEAD": {
    "count": 0
  },
  "GET": {
    "count": 2,
    "bytes": 403064,
    "ranges": [
      "0-32768",
      "3622-253242",
      "460080-580753"
    ]
  },
  "Timing": 1.6177599430084229,
  "cprofile": [
    "   ncalls  tottime  percall  cumtime  percall filename:lineno(function)",
    "        1    0.749    0.749    0.749    0.749 {get_ranges}",
    "        1    0.592    0.592    0.592    0.592 {get}",
    "        1    0.182    0.182    0.182    0.182 {method 'bytes' of 'builtins.GetResult' objects}",
    "        1    0.028    0.028    0.780    0.780 reader.py:86(read)",
    "        1    0.019    0.019    0.807    0.807 __init__.py:99(open)",
    "        8    0.007    0.001    0.007    0.001 {method 'execute' of 'sqlite3.Connection' objects}",
    "        1    0.004    0.004    0.005    0.005 warp.py:453(calculate_default_transform)",
    "        1    0.003    0.003    0.003    0.003 warp.py:113(transform_bounds)",
    "        1    0.002    0.002    0.002    0.002 {built-in method from_url}",
    "      155    0.001    0.000    0.003    0.000 __init__.py:298(__init__)",
    "        1    0.001    0.001    0.001    0.001 core.py:3521(__setmask__)",
    "        1    0.001    0.001    0.009    0.009 rasterio.py:94(__attrs_post_init__)",
    "      148    0.001    0.000    0.010    0.000 base.py:161(read)",
    "        2    0.001    0.000    0.001    0.000 {built-in method posix.mkdir}"
  ]
}
```
😍 

that's said https://github.com/vincentsarago/vsifile/pull/18/files#diff-794a5ccf02e5438e3a4dcb382f92c0ffa92cce74a283ae56fd737ba841bfc217R85 seems to take quite some time :-(